+++ backend/src/books/books.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Book, BookCopy } from '../entities';
+import { Book, BookCopy } from '../database/entities';
 import { PaginationOptionsDto } from '../common/dtos/page-options.dto';
 import {
   CreateBookCopyRequestDto,
